Move render-related files to core module
