+++ src/animations/start_computer_bios.py
@@ -24,8 +24,8 @@ This is the boot animation from the first computer
 # init colors
 import curses
 
-from src.core.render.boot_animation import StyledText, Step, Stage, BootAnimation
-from src.core.render.render import CursesRenderer
+from src.core.boot_animation import StyledText, Step, Stage, BootAnimation
+from src.core.render import CursesRenderer
 
 
 def create_animation(renderer: CursesRenderer) -> BootAnimation:
